Reduce ambiguity of grammar
The `expressions` and `keywords` productions were causing extra
ambiguity in the grammar. This is because I was allowing the keywords
and expressions to be combined into a list. This causes the parser to
make variable length lists of all possible combinations before arriving
at a final production.

While this is inefficient it does produce the correct result. I may move
on from here and come back to revisit it in the future.

To address the failing specs, I use the `.at_least(1).time` chain on
`have_received`. This prevents the test from failing if the query is
instantiated multiple times.