Reduce ambiguity via SPACE associativity
Thanks to @georgebrock and his suggestion to assign the associativity of
the `SPACE` token. This further reduces the ambiguity of the grammar,
and causes the `Query` object to only be created once!

I don't really understand what's happening here, but I'm sort of
thinking of it as telling the parser it can only add items to the end of
the list, rather than at either end. This makes it so there's only one
way to build up the list (or tree?) of space-separated tokens. At least
that's what I'm telling myself for now.