Introduce Assembler
It felt off to have the extraction of clauses in the Query Definition
object.

This commit introduces an Assembler object that is responsible for
building the clauses from the definition via a set of expressions and
keywords. Keywords are not implemented yet.

The QueryDefinition object is now a simple value object, which will
house the list of filters and perhaps other data.