Rename Expression to Query::Input
The over use of the word expression was getting really confusing.

This change replaces the Yuriita::Expression class with
Yuriita::Query::Input. It also creates Input objects for the keyword
scopes instead of using a hash.

There is a little bit of coupling when creating the Query::Input for a
keyword scope, because we have to set the `qualifier` on the Input.
Because the Input requires a `qualifier`, but the qualifier for a
keyword scope is currently implicit ("in") end up with a "magic" string
"in" in the parser. Maybe in the future we can somehow fix this.