Multiple matching filters
Refactored our three separate filters into a single one. This simplifies
the behavior so we can consistently AND or OR clauses together.

As a result of this, the builders no longer make sense, so I removed
them. I plan to add them back in the future once we figure out the right
abstractions. The abstractions may have more to do with the behavior of
multiple matches and how they are combined (or omitted) instead of how
we specify the conditions.