Handle keyword searches
Using the And and Or combinations we can build a query to search for the
given keywords in the provided scopes.

A developer will specify a number of KeywordSearch's. Each of these will
specify how to search for a keyword, such as using an `ILIKE` query.

If the user does not specify a scope, such as `in:title` then all
keyword searches will be used. If they do specify a scope, then only
those scopes will be searched.

When building the KeywordSearch, the developer can specify if the
keywords should be AND'd together or OR'd together. This will determine
if ALL keywords must exist in a scope, or if ANY of the keywords must
exist in a scope.

At the moment, all keyword searches will be OR'd together after ANDing
or ORing the keywords.

If the developer chooses to AND they keywords together, then something
like "cats ducks in:title in:body" will require that the terms "cats"
and the term "ducks" are present in either the title or the body.

In the future it might be nice to provide some configuration to the
developer to allow them to specify if keyword searches should be OR'd or
AND'd. We should provide a default for this configuration.