Fix RSpec and rake interaction
rspec-rails was not in the development group, which required the
addition of

```ruby
require "rspec/core/rake_task"
```

to the Rakefile. This failed on Heroku because that file wasn't
available, because in production rspec-rails isn't included.

This updates the Rake file to not include that file, but to check if
RSpec is defined instead.