Don't use `git` to determine gemspec files
After deploying to Heroku I noticed a number of warnings:

```
app[web.1]: fatal: not a git repository (or any parent up to mount point /)
app[web.1]: Stopping at filesystem boundary (GIT_DISCOVERY_ACROSS_FILESYSTEM not set).
```

After some quick Googling, I found [this great write up from
@schneems][1].

The key takeaway is that Heroku strips the `.git` directory from the
slug in order to keep it small, which means that any calls to `git`
result in the above warning.

So, Who Called Git? I searched the codebase for `git` and found that the
.gemspec file calls `git ls-files -z` to determine the files that should
be included in the gem. I confirmed this call was in fact the culprit
using the method outlined in the blog post.

This commit changes from using `git ls-files -z` to using `Dir.[]`. This
ends up removing a number of files that didn't need to be in the gem
anyway.

Thank you Richard for the detailed post on how you debugged the issue,
you saved me a TON of time.

[1]: https://schneems.com/2016/11/28/who-called-git-an-unusual-debugging-story/