Sync 100 top rated movies via db:seed
We would like a way to consistently update a list of movies in the demo
app.

This commit updates the `db:seed` rake task to delete all current
movies, and then fetch the 100 top rated movies from The Movie Database
API. This requires the `TMDB_API_KEY` environment variable.

We can set up a scheduled task on Heroku to run this periodically to
re-sync data to the demo app.