Change movies revenue column limit from 4 to 8
When importing movie details the revenue for Avengers: Endgame was
returned as `2797800564`. When attempting to save this an exception was
raised:

```
ActiveModel::RangeError: 2797800564 is out of range for
ActiveModel::Type::Integer with limit 4 bytes
```

The default :limit for an integer column is 4 bytes. [This represents
the range from -2147483648 to +2147483647][1].

This commit update the limit of the column to 8 which will allow for
larger integers.

[1]: https://www.postgresql.org/docs/12/datatype-numeric.html