Combine all inputs into a single list.
Because we handle all matching at the filter level, we don't need to
preemptively split the "types" of inputs into different lists.

This consolidates all the inputs into a single list in the query. When
filters are run, all inputs are passed to it and the filter's matchers
select the appropriate inputs.