Extract QueryBuilder
The `Yuriita.sift` method delegated almost directly to the Runner class.
The Runner class was responsible for taking a string input and returning
either a success or error Result. This API was a bit too fixed. There is
no way for a caller to decide to ignore parse errors and simply return a
relation, and there's no way to do any validation of the input before
attempting to run it.

This commit extracts a QueryBuilder object from the Runner class. The
QueryBuilder now takes a string input and returns a Query on success or
raises a Yuriita::ParseError on failure. This error can be rescued by
the caller if they wish to ignore it.

The Runner now accepts a Query object and always returns a relation.
This makes it easier to compose the behavior you wish to have.