Only allow a single matcher
The idea of multiple matchers for a single filter was a bit half baked.
The idea was that you could define a filter such that it would match
either of these:

```
is:published
state:published
```

While I do think this feature is valuable, I don't think we're ready to
tackle it. There are a number of open questions that haven't been solved
yet:

- For a filter that matches multiple inputs, how do you specify the
canonical form to use when building a new query string?
- If multiple matching inputs are present for a single filter, how do
those inputs change when switching to a different menu item within the
same menu? If the current input is
  ````
  state:published is:published
  ````
  and the user selects "Draft" from the UI what is the new input string?