Remove negation from expression filtering
When defining an expression filter, we would like to allow a developer
to act on a relation directly, rather than provide a set of conditions.

We'd like to change this

```ruby
Yuriita::ExpressionFilter.new(options) do |term|
  { published: true }
end
```

into this

```ruby
Yuriita::ExpressionFilter.new(options) do |relation|
  relation.published
end
```

This allows us to re-use scopes defined on the relation rather than
having to redefine them ourselves. It also solves the `joins` problem by
allowing a developer to do the join themselves if they need to.

The downside is there is no way to do automatic negation. The plan is to
allow a developer to define negation by means of another block. This
will require a large refactor to use relations instead of conditions,
but to prepare for that we're removing negation now.