Use Zeitwerk for code loading
[Zeitwerk] is the new code loader in Rails 6 and provides a reasonable
approach to file structure and constant loading. It removes the need to
add `require` statements and ensure the ordering of those statements is
correct.

This commit adds the Zeitwerk code loading and configures it to eager
load the entire gem.

If someone adds `gem "yuri-ita"` to their Gemfile Bundler will load
`yuri-ita.rb`. This file is set up to `require "yuriita.rb"` so we don't
have to deal with inflections. We ignore the `yuri-ita.rb` file and rely
on `yuriita.rb` to be the main entry point.

I also removed a number of places where we were calling

```
require "spec_helper"
```

This is done by `.rspec` and is unnecessary.

[Zeitwerk]: https://github.com/fxn/zeitwerk