Add Enumerable to Query
We would like to be able to iterate over the inputs in a query so we can
format each in turn. We would also like to be able to add inputs or
remove inputs matching a certain condition.

This commit add the Enumerable module to Query. It adds methods to add
and delete inputs, as well a `each_element` which will yield all the
inputs and then all the keywords. In the future when we figure out how
to preserve the ordering of inputs and keywords `each_element` should
yield items in their original order.