Use formatter in collections to output new query
Each collection generates a list of options for display. Each option
needs to have a query string that will be used to apply (or unapply)
that option.

The formatter is responsible for taking a query and turning it into a
params hash. The hash contains a single element: the `param_key` whose
value is the new input string.

This commit updates the collections to accept a formatter which it uses
to build the new params.

There's one oddity about this that I hope to address in the near future.
Each collection is used for two purposes. The first is for running the
actual query. This is done via the `apply` method. The second is
generating the view options. This is done via the `view_options` method.

Only the view options requires the formatter. When we build a collection
to apply the query, we don't know the param key, and thus we cannot
build a formatter. So we can't pass a formatter to the collection in the
definition's `apply` method. Because of this, each collection defaults
the `formatter` to `nil`.

This highlights the idea that we really want two separate objects here,
one to handle applying a definition's filters, and another object for
generating view options from a definition. This can only happen once we
have a common interface for both objects, which we don't have at
the moment. Future efforts will work toward this goal.