Remove expression matcher
Having a notion of matching was making it hard to understand when to add
or remove inputs from the query when building the params.

This commit removes the expression filter and search filter `matcher`
argument. It's replaced with an `input` argument. This input is used to
determine if an option is selected. It's also used to generate the new
query when an option is selected.