Extract selector objects from collections
This commit refactors the collection options to extract the selection
behavior for each of them.

This commit also fixes a TODO item which was to make Option equality
based on name rather than object identity. This allows us to define an
exexclusive collection with a default, and if no option is selected via
an input the default option is selected.