Combine inputs and keywords to maintain order
Splitting a query's terms into inputs and keywords makes it difficult to
maintain the original order when turning a query back into a string.  In
investigating this, it became clear that the distinction between
keywords and inputs was only necessary when performing the search. When
searching, keywords can be separated from inputs by checking to see if
an item is an input object, then only searching on items that are not
(we could also check to see if it is a string, and only search the
strings).

This combines keywords and inputs into one array in a query or
configuration. This maintains the order, so when a query is converted
back to a string, the user sees the string they originally submitted.