Load factories in library specs
When running specs that do not require the rails app to be loaded, only
the `spec_helper` file is required. This file is required automatically
by RSpec via the `.rspec` configuration file.

FactoryBot does not automatically load factory definitions without
Rails. This commit ensures that the definitions are loaded even when the
Rails app isn't booted so we can access factories that build library
objects such as inputs, options, filters, and queries.