Extract parameters objects for building params
When a user selects or deselects an option we navigate to a new URL to
apply or un-apply that option. This requires that we be able to generate
the new parameters for the query based on the behavior of the collection
of options.

There are two behaviors for options, Single and Multi. Multi parameters
either add or remove the option in question from the params. Single
parameters replace all existing inputs in the collection with the option
in question on selection, or remove itself on de-selection.

This commit extracts these behaviors from the Collection objects into
their own classes. The collection objects are very slim now and in a
future commit could likely be removed and have their logic moved into
the Definition classes.