Extract Clauses from Collections
As part of the work to split the responsibilities of the xxxCollection
objects into two separate concerns, this commit extracts the behavior of
applying the selected filters to the relation.

Two "Clauses" were extracted, one for filters and one for search. The
only real difference is that the search clause needs the keywords so it
can pass the keywords to the filter when it applies it.

Both clauses take a combination object with which they use to combine
multiple filters. In the case of the Single and Exclusive definitions,
there will always only be a single filter, but we pass a combination so
that those definitions can conform to the same interface. Combining a
single relation is effectively a no-op, so this is safe to do.