Add means to check if table is filtered
We would like to allow users to "reset" their table to some default
state. Instead of providing an explicit means to reset, we instead
provide a way to check if the table is currently filtered. By doing so
we can leave what it means to reset the table to the developer or
designer.