Create initial layout for movies demo (#81)
This adds several styles to the demo app, arranging it into a 2-column
layout with the filters and sorting options to the right and the movie
list on the left. The styles don't indicate the art direction of the app,
rather this just creates a scaffolding to work from.

This removes items which are no longer used, such as the index table and
toggle group.