Refactor input parsing to remove Fragment
Originally, our inputs were split into various groups within the Query
object. We separated keywords, expression inputs, and search scopes into
different internal data structures in the Query. We then realized that
in order to preserve the input ordering when we displayed the query back
to the user, it was easiest to simply store an array of the inputs in
the Query.

The Fragment object represented a portion of the full query as we were
building it up. It knew how to appropriately merge itself with another
fragment to maintain the separation of the keywords, expressions, and
scopes. Now that our inputs are stored in a simple array we no longer
need the Fragment object.

This commit  removes the Fragment and reorganizes the parser
productions. The `fragment` production has been split into two
productions, `input` and `input_list`. The `input` production now
captures all the possible input types. This makes it clear what the
various forms of our inputs are. The `input_list` production is either a
single input, or a list of inputs followed by a single input.