Add yarn binstub in example app
Webpack compilation was failing on Heroku. After a bunch of fiddling I
found that adding a binstub for yarn in the example app's bin directory
fixed the issue. My guess is that as part of webpack compilation, it
attempts to compile from the Rails root directory not the project root
directory.

Adding a duplicate yarn binstub appears to fix the issue. It's not
ideal, but this is just a demo app and isn't really intended to be
production level quality.

When adding new packages or installing existing ones, you can continue
to use `bin/yarn` from the project root.