Split inputs into distinct classes
In order to support numeric inputs with an operator such as

```
rating:>7.0
```

we need to be able to capture the operator in some way. This operator
does not have a place to live in the Input class, nor would it really be
proper to add an optional operator to all inputs.

This commit splits all the various input types into their own classes.
This will allow us to capture additional data for particular inputs such
as the operator on numeric inputs, or negation when we support it.

Now that inputs have their own type, we can create factories for each
input type. Most of the changes in this PR are updates to the specs to
use the new factories.