Remove overview from search
It was confusing to have the overview be a default search field in the
demo because if you search for something like "seven", a number of
movies will show up but it won't be clear why they were in the results
as the overview field is not displayed in the list.

This removes the overview field from the search definition.