Add Author to Posts
We'd like to create a "dynamic filter" that allows a user to input a
search like

```
author:eebs
```

and define a filter that would take the term from the input, and use
that as part of the search.

This adds authors to posts in our example app, which will be used to
build the dynamic filter feature.