Add Yuriita::Search to allow custom param_keys (#93)
This allows a user to define any param_key value and then call
table.search.x, where x is the param_key, in order to retrieve the input
string. The search object was necessary in order to avoid conflicts if
the user defines param_key to be the name of a method that already
exists on Yuriita::Table.