Add Clauses::Identity
As part of execution, all Definitions are given the Query and need to
provide a Clause object back. This Clause is applied to the initial
Relation. This Clause object updates the relation to change the database
query. When a Definition does not match the Query, it still must return
a Clause.

Previously, we would return a Filter or Search clause. These clauses
would check their internal state, and if no filters or keywords were
present they would return the relation unchanged.

This commit introduces an Identity clause that can be returned from the
definition when it does not match the query. This allows the Filter and
Search clauses assume they will only be used when they actually have
something to do.