Organize classes (#94)
As the codebase evolves various organizational structures are emerging.

It's clear we will have a number of Definition objects that effectively define different sets of behavior. They have a well-defined interface that allows for flexible behavior. They are similar to the adapter pattern.

We also have a number of Clause objects that represent various bits of logic, defining how a relation is filtered by a set of matching inputs.

There is also the concept of Selects, which are responsible for determining which filters are selected given the query.

Creating an organizational structure for these classes through namespacing will help to group related behavior, and make the organization of the codebase clearer.