Use `form_with` for search input so params are sent correctly
When we added the Search object to the Table, it broke the search input
in the demo application. When using
`form_for(table.search)`
it changes the parameters that are submitted to the server to
`{"yuriita_search"=>{"query"=>"king"}}`

This uses `form_with` instead of `form_for` so that the input params are
not wrapped in the model name.