Words can contain any character.
This commit expands the lexer's definition of a word to include any
character that doesn't already have some other special meaning within the
grammar (i.e. whitespace, `:` and `"`).

This allows Yuri-ita to be used for a greater variety of situations, for
example filtering by users' email address or by customers' order numbers.