Remove special casing for sort and scope inputs
There is no need to parse `in:title` and `sort:created-asc` as special
inputs. These can be expressed as Expression inputs as they follow the
`qualifier:term` pattern. They were overly restrictive in that you could
not use `as:author` to define a scope to search keywords. For example,
it may be desirable to create a configuration that caused

```
james as:author
```

to search for the keyword `james` as the author of a post. Previously,
you would be forced to use `in:author` which may not be as clear.

The same applies to sort inputs. You can now use `order:title` if you
wanted to.

*This change requires that all configuration options update any scope or
sort inputs to use the expression input*

```diff
- Yuriita::Inputs::Scope.new("in", "title")
+ Yuriita::Inputs::Expression.new("in", "title")
```

```diff
- Yuriita::Inputs::Sort.new("sort", "title-asc")
+ Yuriita::Inputs::Expression.new("sort", "title-asc")
```